requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
codaio==0.6.12
schedule==1.2.2
gunicorn==21.2.0
